# src/swagger.yaml
swagger: "2.0"
info:
  version: "1.0"
  title: "MyMediathek"
  description: "Interface to synchronize and provide bookmarked movies and handle movies."
basePath: "/api"
...
schemes:
- "http"
- "https"
paths:
...
  /info/config:
    patch:
      tags:
      - "Info"
      summary: "Change config data on server"
      operationId: "server.interface.info_config_patch"
      consumes:
      - application/json
      parameters:
      - in: "body"
        name: "config"
        required: true
        schema:
          type: "object"
          properties:
            name:
              type: "string"
              description: "A valid config parameter name"
              enum: [minMovieStateAge, expiryBookmarkWarningDays, expiryMovieWarningDays, searchChannelInclude, searchTitleFilter, searchTopicFilter]
            value:
              type: "string"
      responses:
        "204":
          description: "successful operation"
  /info/statistics:
    get:
      tags:
      - "Info"
      summary: "Get Server statistics"
      description: "Get details of stored data"
      operationId: "server.interface.info_statistics_get"
      parameters:
      - name: "daysbefore"
        in: "query"
        description: "Number of days before current date for moviestate calculation."
        type: "integer"
      produces:
      - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "object"
            properties:
              tables:
                type: "array"
                items:
                  type: "object"
                  properties:
                    name:
                      type: "string"
                    nb:
                      type: "integer"
              bookmarks:
                type: "object"
                properties:
                  number:
                    type: integer
                  station:
                    type: "array"
                    items:
                      type: "object"
                      properties:
                        name:
                          type: "string"
                        nb:
                          type: "integer"
                  categories:
                    type: "array"
                    items:
                      type: "object"
                      properties:
                        name:
                          type: "string"
                        nb:
                          type: "integer"
              moviestate:
                type: "object"
                properties:
                  bookmarked:
                    type: integer
                  bookmarkedactual:
                    type: integer
                  bookmarkedold:
                    type: integer
                  oldnotbookmarked:
                    type: integer
                  oldtotal:
                    type: integer
                  total:
                    type: integer
...
  /moviestate:
    get:
      tags:
        - "Moviestate"
      summary: Returns a list of moviestates
      description: Returns moviestate<p>Optional filter parameter "before" can be used to return onlymoviestates older than the given date time stamp.<p>
      operationId: server.interface.movieStateGet
      parameters:
        - name: before
          in: query
          description: timestamp as filter
          type: integer
        - name: "limit"
          in: "query"
          description: "number of records to return"
          type: "integer"
        - name: "offset"
          in: "query"
          description: "offset for first value to return "
          type: "integer"
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: "array"
            items:
              $ref: "#/definitions/moviestate"

    post:
      tags:
      - "Moviestate"
      summary: "Get multiple states with ids"
      operationId: "server.interface.movieStatePost"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "ids"
        description: "Array of ids to be querried."
        required: true
        schema:
          type: "array"
          items:
            type: "integer"
            format: "int64"
          example: [234323311, 456456454, 774343234]
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
             $ref: "#/definitions/moviestate"
        "400":
          description: "Invalid data"

    delete:
      tags:
      - "Moviestate"
      summary: "Delete movie states older than a given days"
      description: "Delete movie states older days before current date. If the difference to today is less than 14 days the force flag must be set in the query."
      operationId: "server.interface.movieStateDelete"
      produces:
      - "application/json"
      parameters:
      - name: "daysbefore"
        in: "query"
        description: "Number of days before current date. If ommited a default from 183 days is assumed."
        type: "integer"
      responses:
        "200":
          description: "successful operation
          schema:
            $ref: "#/definitions/moviestate"
  /moviestate/{hid}:
    get:
      tags:
        - "Moviestate"
      summary: Return a movie state.
      description: Returns the current state for the given id.
      operationId: server.interface.movieStateIdGet
      parameters:
      - name: "hid"
        in: "path"
        description: "Hash ID of the movie"
        required: true
        type: "integer"
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            $ref: "#/definitions/moviestate"
        '404':
          description: "id not found"

    patch:
      tags:
      - "Moviestate"
      summary: "Update/create state of movie"
      description: "Updates or creates a new entry in moviestate list for the given id, always returns success."
      operationId: "server.interface.movieStateIdPatch"
      consumes:
      - "application/json"
      parameters:
      - name: "hid"
        in: "path"
        description: "Hash ID of the movie to be updated/created in the movie state table"
        required: true
        type: "integer"
        format: "int64"
      - name: "seen"
        in: "query"
        description: "Seen state of the movie"
        type: "boolean"
        default: false
      responses:
        "200":
          description: "successful operation
          schema:
            $ref: "#/definitions/moviestate"
  /categories:
    get:
      tags:
        - "Categories"
      summary: Returns a list of category names with the number of entries
      description: Returns the names of available categories with the number of entries in the book mark list.<p>Optional filter parameter min or max number can be used to return only categories with at least min resp. max number of entries.<p>
        <b>Note Only one filter can be applied in one call!</b><p>Returns an empty list if no (matching) categories are found.
      operationId: server.interface.categories_get
      parameters:
        - name: minNb
          in: query
          description: min number of entries as filter
          type: integer
        - name: maxNb
          in: query
          description: max number of entries as filter
          type: integer
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              type: object
              properties:
                name:
                  type: string
                nb:
                  type: integer
  /categories/{name}:
    delete:
      tags:
        - "Categories"
      summary: Delete given category
      description: ''
      operationId: server.interface.categories_byname_delete
      parameters:
        - name: name
          in: path
          description: name of the category to delete
          required: true
          type: string
      responses:
        '200':
          description: Successful operation
        '403':
          description: Category is in use
        '404':
          description: Category not found
  /player:
    post:
      tags:
      - Player
      summary: Add new player
      description: Add a new or modify an existing player.
      operationId: server.interface.player_new
      consumes:
      - application/json
      parameters:
      - in: "body"
        name: "player"
        description: "Definition of player to added"
        required: true
        schema:
          type: "object"
          required:
            - address
            - port
          properties:
            address:
              type: "string"
              description: Either an IP address or a valid netowrk name like 'localhost'
            port:
              type: "integer"
            name:
              type: "string"
              description: If name is ommited, 'address:port' is used as name.
            type:
              type: "string"
            authentification:
              type: "string"
              description: "Authentification parameters to be used. Attention these parameters are visible in the network"
            default:
              type: "boolean"
              description: "Mark this player as default. Note: The first player created will always be marked as default, even if this flag is not set."

          example: {"address": "localhost", "port": 8090, name: "KODI lokal", "type": "kodi"}
      responses:
        "204":
          description: "successful operation"
        "400":
          description: "Invalid request, e.g. body syntax wrong"
        "422":
          description: "Invalid data"
  /player/{playerid}:
    delete:
      tags:
        - Player
      summary: Delete given player
      description: ''
      operationId: server.interface.player_delete
      parameters:
        - name: playerid
          in: path
          description: Id of player to be deleted
          required: true
          type: string
      responses:
        '200':
          description: Successful operation
        '404':
          description: Player not found
  /player/{playerid}/test:
    get:
      tags:
        - Player
      summary: test the player connection and return result
      description: ''
      operationId: server.interface.player_test
      parameters:
        - name: playerid
          in: path
          description: Id of player
          required: true
          type: string
      responses:
        '200':
          description: Test was executed, result returned in JSON
          schema:
            type: "object"
            properties:
              success:
                type: boolean
                description: "True if player could be successfully reached, False if an error occured"
              detail:
                type: "string"
                description: "Explanation of status"
            example:
             {"success" : False, "detail": "No connection to player" }
        '404':
          description: Player ID not found
  /player/play:
    post:
      tags:
      - Player
      summary: Play URL
      operationId: server.interface.player_play
      consumes:
      - application/json
      parameters:
      - name: playerid
        in: query
        type: string
      - name: movie"
        in: body
        description: Movie address to be played
        schema:
          type: object
          required:
            - url
          properties:
            url:
              type: string
            title:
              type: string
            timestamp:
              type: "integer"
              description: "Start time stamp in seconds since since Jan 01 1970. Used to prevent double starts. If time stamp before current time, request will be denied wit 425"
      responses:
        200:
          description: successful operation
        400:
          description: Invalid data
        425:
          description: (Assumed)) Double start denied
        500:
          description: player error, e.g. no player configured
definitions:
...
  moviestate:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "Java hash value of sender and url:<br>(java) int id = (sender + url).hashCode().abs()"
      seen:
        type: "boolean"
        default: false
        description: "indicates if the movie is seen"
      bookmarked:
        type: "boolean"
        default: false
        description: "indicates if the movie is in the bookmark table"
      expiry:
        type: "integer"
        description: "Expiry time stamp in seconds since since Jan 01 1970. If no expiry is set the value is 5000000000."
    example: {"id" : 234323311, "seen" : true, "bookmarked": false, "expiry" : 1636758000 }
...